Transana 2.99a5 - Update 12
On OS X 10.7.5, my RichTextCtrl becomes unstable with wxPython 3.0.2.0.
It becomes easy to type faster than characters are displayed, and rather
than buffering the characters, characters just get dropped.  This
doesn't appear to be the case on OS X 10.10.3.  So in reverting
wxPython, the Toolbar for the Document window shifts from (16, 16) to
(32, 32).  These changes try to make that toolbar look and fit better.